python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "protobuf_experiments"
version = "0.1.0"
dependencies = [
    "protobuf",
    "websockets",
    "PyYAML",
    "aiosqlite",
    "python-telegram-bot",
]

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
